architecture refactor
