--- internal/repository/ticketRepository.go
@@ -1,3 +1,0 @@
-package repository
-
-var TicketList []domain.Ticket
