+++ internal/tickets/repository/ticketRepository.go
@@ -1,0 +1,5 @@
+package repository
+
+import "github.com/ZoeAgustinaTira/goBases-Airline/domain"
+
+var TicketList []domain.Ticket
